fix: Allow unset of groupings by storing as empty string.
To unset the groupings attribute, we pass an empty string as the value
in the API request. The API then responds with a payload that has a
null value for the groupings.

There is a known issue in the provider framework in this situation:
https://discuss.hashicorp.com/t/after-applying-this-test-step-and-performing-a-terraform-refresh-the-plan-was-not-empty/42835

The hack solution is to store the groupings value as an empty string
in terraform state when it has been removed.